+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 * Reset resolved `owner` in context on component remove
 * Increase size limit to `2.5 KB`
+* Replace `unbind` with `onRemove` 
 
 ## 6.0.0-alpha.2 
 * Fix typo (`setupForComponet` -> `setupForComponent`)